+++ src/backend/resolvers/mutations/organization/index.js
@@ -2,6 +2,7 @@ const providerOrganizationResolvers = require('./provider')
 const payerOrganizationResolvers = require('./payer')
 const pathwaysOrganizationResolvers = require('./pathways')
 const apmOrganizationResolvers = require('./apm')
+const obmOrganizationResolvers = require('./obm')
 const metaResolvers = require('./meta')
 
 module.exports = {
@@ -10,4 +11,5 @@ module.exports = {
   ...payerOrganizationResolvers,
   ...pathwaysOrganizationResolvers,
   ...apmOrganizationResolvers,
+  ...obmOrganizationResolvers,
 }