+++ src/backend/typeDefs/mutations/organization/index.js
@@ -2,6 +2,7 @@ const providerTypeDefs = require('./provider')
 const payerTypeDefs = require('./payer')
 const pathwaysTypeDefs = require('./pathways')
 const apmTypeDefs = require('./apm')
+const obmTypeDefs = require('./obm')
 const metaTypeDefs = require('./meta')
 
 module.exports = [
@@ -9,5 +10,6 @@ module.exports = [
   ...payerTypeDefs,
   ...pathwaysTypeDefs,
   ...apmTypeDefs,
+  ...obmTypeDefs,
   ...metaTypeDefs,
 ]