+++ src/backend/typeDefs/mutations/organization/obm/create.js
@@ -1,0 +1,22 @@
+const { gql } = require('apollo-server-express')
+
+const createObmOrganizationTypeDefs = gql`
+  input CreateObmOrganizationInput {
+    slug: String!
+    organization: String
+    organizationTiny: String
+    businessModel: String
+    start: Int
+  }
+
+  type CreateObmOrganizationPayload {
+    _id: ID!
+    slug: String!
+    organization: String
+    organizationTiny: String
+    businessModel: String
+    start: Int
+  }
+`
+
+module.exports = createObmOrganizationTypeDefs