+++ src/frontend/Orion/Query/NewQueryTool/PqlView.js
@@ -1,21 +1,9 @@
 import React from 'react'
-import { Link } from 'react-router-dom'
-import Inspector from 'react-inspector'
-import styled from '@emotion/styled'
 
-import Spinner from 'frontend/components/Spinner'
-
-import Spacing from '../../../utils/spacing'
-import FontSpace from '../../../utils/fontspace'
-import Color from '../../../utils/color'
 import useAquila from '../../../hooks/useAquila'
 import { Colors } from '../../../utils/pulseStyles'
 
-const Wrapper = styled.div({
-  width: '100%',
-  padding: Spacing.S4,
-  background: Color.LIGHT_BLUE_GRAY_1,
-})
+import QueryToolTable from './QueryToolTable'
 
 const PqlView = () => {
   const {
@@ -25,12 +13,10 @@ const PqlView = () => {
     submitPql,
   } = useAquila()
 
+  const businessObjectName = pql.match(/[\w\s]+={.*}/) && pql.match(/[\w\s]+=/)[0].replace('=', '')
+
   return (
-    <Wrapper>
-      <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-        <h1 style={{ padding: Spacing.S4, ...FontSpace.FS4 }}>Query Tool</h1>
-        <Link style={{ fontWeight: 700, padding: 12, background: 'blue', color: 'white', borderRadius: 4 }} to="/orion/query/tool/placard">Placard View</Link>
-      </div>
+    <>
       <form style={{ margin: 12, boxSizing: 'border-box' }}>
         <input
           style={{ width: '94%', height: 20, padding: 12, background: Colors.WHITE, border: '1px solid black' }}
@@ -49,20 +35,15 @@ const PqlView = () => {
         </button>
       </form>
       {
-        loading && <Spinner />
-      }
-      {
         results.error && <span style={{ color: 'red' }}>{results.error}</span>
       }
-      <div style={{ padding: 12 }}>
-        <h2>Results (Randomized Sample -- Limit 50)</h2>
-        <Inspector
-          table
-          theme="chromeDark"
-          data={results.error ? [] : results}
-        />
-      </div>
-    </Wrapper>
+      <QueryToolTable
+        data={results}
+        loading={loading}
+        businessObjectName={businessObjectName}
+        afterMutationHook={() => submitPql(pql)}
+      />
+    </>
   )
 }